refactor: extract handling catch event migration methods into a class
Handling catch events logic grown as we add more supported catch event types. That lead to
creation of many methods accordingly and overwhelmed the actual migration processor. A new class is
created to apply handling catch event migrations. That way, the main migration processor class will
focus on main migration parts whereas the newly created class will handle catch event migrations.
